Refactor User as a record
The more boilerplate code I added to the User class the more I wondered
if it was the right approach, and if there was a simpler approach to
take. It felt increasingly bloated and wrong. Gladly, while refactoring
the class to use Lombok, I learned about records in Java 14. From what I
read on https://www.baeldung.com/java-record-keyword, they're the
approach that makes sense, exactly what I'm looking for.

Given that, this change refactors the class to become a record.