Finish implementing login functionality
While a bit rough, this is the first implementation of the login
functionality, where the user can both load the login form and submit
it, whereupon, if their username is found and from it their phone
number, a verification code will be sent via SMS. As I said, it's a
little rough, but it works. A refined implementation will be added
shortly.