Finish implementing the verifyme route
This is a bit of a clunky commit, but provides all of the required
changes. It adds a post verifyme endpoint that processes requests to
verify the verification code sent via SMS to the user. It also extracts
some of the existing functionality out into a new UserService class,
with the intention of making the code more maintainable and readable.
Otherwise, even though this code will only be used in a tutorial (I
expect), it was still getting unwieldly.